CalendarView header not updated on year change.
The calendar view was updating its header only if the focused
month changes. Removed the check whether the displayed month
changed before setting it since the setter is carefully called
only in cases when there is a change. Hence, now if the year
or the month change, we update the header. This is the safest
and least intrusive fix.

bug:9167305

Change-Id: I98200bb57580f6416abd30c6c25454d4474add64